requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
urllib3==2.2.1
Werkzeug==3.0.3